Bust Docker cache on nginx.conf.tmpl modification.
Using ADD on a directory causes Docker builds to skip any attempt at
checking if directory contents were modified since the last run.

Since we want to bust cache when we modifiy the configuration,
explicitly define the configuration file to allow Docker to check if the
file was modified.

We may need to do this for sbin files in the future, too.